Input system
Key presses are passed from UI to input system through registerKeyPress
in normal, visual and command modes. Custom mode is used for picker and
grepper (until I figure out better way to model them) and in this mode
UI handles input directly. It is now possible to create true key
mapping, i.e. map some key sequence to another sequence, instead of only
mapping of some key sequence to command. Mapping is done per mode
(normal, visual, command).

Command line has been split between core and UI. Core's CommandLine
has its own key handling to implement readline-like functionality
instead of relying on FTXUI's Input event handling. FTXUI's Input is used
only to render the command line.

Others:
* Files is now a struct (so that it would be easy to do forward
  declaration without including <list>); list contains a std::list<File>
  instead of previously used std::list<std::unique_ptr<File>>
* move Severity definition to severity.hpp so that user_interface.hpp
  will no longer include logger.hpp
* display input state on the right side of status line
* commands picker puts selected command name in command line on accept

+++ src/ui/main_view.hpp
@@ -19,7 +19,7 @@ struct MainView final : UIComponent
     ftxui::Element render(core::Context& context) override;
     void reload(Ftxui& ui, core::Context& context);
     ViewNode& createView(std::string name, ViewNode* parent = nullptr);
-    void scrollTo(Ftxui& ui, ssize_t lineNumber, core::Context& context);
+    void scrollTo(Ftxui& ui, long lineNumber, core::Context& context);
     const char* activeFileName() const;
     bool isViewLoaded() const;
     View* currentView();
